# openapi/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "77-Bit API",
    "description": "An API for interacting with the 77-Bit L2 blockchain.",
    "termsOfService": "https://77-bit.com/terms",
    "license": {
      "name": "License",
      "url": "https://77-bit.com/license"
    },
    "version": "1.0",
    "x-logo": {
      "url": "https://pbs.twimg.com/profile_images/1642220143915696128/x3yCWwbE_400x400.jpg",
      "altText": "77-Bit logo"
    }
  },
  "servers": [
    {
      "url": "https://testnet-api.77-bit.com",
      "description": "Testnet server (Goerli)."
    },
    {
      "url": "https://api.77-bit.com",
      "description": "Mainnet server (Ethereum)."
    }
  ],
  "paths": {
...
    "/api/v1/transfers": {
      "post": {
        "tags": [
          "Transfer"
        ],
        "summary": "Transfer Asset",
        "description": "This endpoint allows for transferring assets between users.",
        "operationId": "Transfer",
        "requestBody": {
          "description": "The transfer request.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Returns the vaults updated by the transfer operation.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VaultDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "The transfer request was invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden."
          }
        },
        "security": [
          {
            "oauth2": [
              "write:transfers"
            ]
          }
        ]
      }
    },
    "/api/v1/transfers/signable": {
      "get": {
        "tags": [
          "Transfer"
        ],
        "summary": "Get Signable Details",
        "description": "This endpoint allows for fetching details of a transfer to be signed",
        "operationId": "TransferSignable",
        "parameters": [
          {
            "name": "from_user_id",
            "in": "query",
            "required": true,
            "schema": {
              "title": "Sender User ID",
              "type": "string",
              "description": "The unique identifier of the user sending the transfer.",
              "format": "uuid"
            }
          },
          {
            "name": "to_user_id",
            "in": "query",
            "required": true,
            "schema": {
              "title": "Receiver User ID",
              "type": "string",
              "description": "The unique identifier of the user receiving the transfer.",
              "format": "uuid"
            }
          },
          {
            "name": "asset_id",
            "in": "query",
            "required": true,
            "schema": {
              "title": "Asset ID",
              "type": "string",
              "description": "The unique identifier of the asset being transferred.",
              "format": "uuid"
            }
          },
          {
            "name": "token_id",
            "in": "query",
            "schema": {
              "title": "Token ID",
              "type": "string",
              "description": "The hexadecimal string representation of the token ID, if applicable (ie. ERC-721/ERC-1155).",
              "format": "hex"
            }
          },
          {
            "name": "amount.IsPowerOfTwo",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "amount.IsZero",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "amount.IsOne",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "amount.IsEven",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "amount.Sign",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the signable transfer details.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransferSignableDto"
                }
              }
            }
          },
          "400": {
            "description": "The signable transfer request was invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden."
          }
        },
        "security": [
          {
            "oauth2": [
              "write:transfers"
            ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DataAvailabilityModes": {
        "enum": [
          "ZkRollup",
          "Validium"
        ],
        "type": "string"
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "SignatureModel": {
        "required": [
          "r",
          "s"
        ],
        "type": "object",
        "properties": {
          "r": {
            "title": "Signature R",
            "minLength": 1,
            "type": "string",
            "description": "The R component of the ECDSA signature, represented as a hexadecimal string.",
            "format": "hex"
          },
          "s": {
            "title": "Signature S",
            "minLength": 1,
            "type": "string",
            "description": "The S component of the ECDSA signature, represented as a hexadecimal string.",
            "format": "hex"
          }
        },
        "additionalProperties": false,
        "description": "Model representing an ECDSA signature data."
      },
...
      "TransferModel": {
        "required": [
          "amount",
          "assetId",
          "expirationTimestamp",
          "fromUserId",
          "nonce",
          "signature",
          "toUserId"
        ],
        "type": "object",
        "properties": {
          "fromUserId": {
            "title": "Sender Vault ID",
            "type": "string",
            "description": "The unique identifier of the transfer sender vault.",
            "format": "uuid"
          },
          "toUserId": {
            "title": "Receiver Vault ID",
            "type": "string",
            "description": "The unique identifier of the transfer recipient vault.",
            "format": "uuid"
          },
          "assetId": {
            "title": "Asset ID",
            "type": "string",
            "description": "The unique identifier of the asset being transferred.",
            "format": "uuid"
          },
          "tokenId": {
            "title": "Token ID",
            "type": "string",
            "description": "The hexadecimal string representation of the token ID, if applicable (ie. ERC-721/ERC-1155).",
            "format": "hex",
            "nullable": true
          },
          "expirationTimestamp": {
            "title": "Expiration Timestamp",
            "type": "integer",
            "description": "The timestamp at which this transfer becomes invalid, in seconds since the Unix epoch.",
            "format": "int64"
          },
          "nonce": {
            "title": "Nonce",
            "maximum": 2147483647,
            "minimum": 0,
            "type": "integer",
            "description": "The unique nonce for the transfer.",
            "format": "int32"
          },
          "amount": {
            "title": "Amount",
            "type": "string",
            "description": "The amount of the asset to be transferred."
          },
          "signature": {
            "$ref": "#/components/schemas/SignatureModel"
          }
        },
        "additionalProperties": false,
        "description": "Request model to transfer assets between users."
      },
      "TransferSignableDto": {
        "type": "object",
        "properties": {
          "signable": {
            "title": "Signable Payload",
            "type": "string",
            "description": "The signable payload for the transfer.",
            "format": "hex",
            "nullable": true
          },
          "expirationTimestamp": {
            "title": "Expiration Timestamp",
            "type": "integer",
            "description": "The timestamp at which this transfer becomes invalid, in seconds since the Unix epoch.",
            "format": "int64"
          },
          "nonce": {
            "title": "Nonce",
            "type": "integer",
            "description": "The unique nonce for the transfer.",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "VaultDto": {
        "type": "object",
        "properties": {
          "vaultId": {
            "title": "Vault ID",
            "type": "string",
            "description": "The ID of the vault.",
            "format": "uuid"
          },
          "vaultChainId": {
            "title": "Vault Chain ID",
            "type": "string",
            "description": "The StarkEx ID of the vault."
          },
          "assetSymbol": {
            "title": "Asset Symbol",
            "type": "string",
            "description": "The symbol of the asset associated with the vault.",
            "format": "string",
            "nullable": true
          },
          "tokenId": {
            "title": "Token ID",
            "type": "string",
            "description": "The token id of the asset associated with the vault, if the asset is an ERC721 or ERC1155.",
            "format": "hex",
            "nullable": true
          },
          "mintingBlob": {
            "title": "Minting Blob",
            "type": "string",
            "description": "The minting blob of the asset associated with the vault, if the asset is a Mintable ERC20, ERC721 or ERC1155.",
            "format": "hex",
            "nullable": true
          },
          "assetStarkExId": {
            "title": "Asset StarkEx ID",
            "type": "string",
            "description": "The StarkEx ID of the asset associated with the vault.",
            "format": "hex",
            "nullable": true
          },
          "userStarkKey": {
            "title": "STARK Key",
            "type": "string",
            "description": "The STARK key of the user associated with the vault.",
            "format": "hex",
            "nullable": true
          },
          "availableBalance": {
            "title": "Available Balance",
            "type": "string",
            "description": "The available balance of the vault."
          },
          "accountingBalance": {
            "title": "Accounting Balance",
            "type": "string",
            "description": "The accounting balance of the vault."
          },
          "dataAvailabilityMode": {
            "$ref": "#/components/schemas/DataAvailabilityModes"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "oauth2": {
        "type": "oauth2",
        "flows": {
          "clientCredentials": {
            "tokenUrl": "https://sevensevenbit-dev.eu.auth0.com/oauth/token",
            "scopes": {
              "write:users": "Access user write operations.",
              "read:users": "Access user read operations.",
              "write:assets": "Access asset write operations.",
              "read:assets": "Access asset read operations.",
              "mint:assets": "Access mint operations.",
              "read:vaults": "Access vault read operations.",
              "write:transfers": "Access transfer write operations.",
              "write:settlements": "Access settlement write operations.",
              "read:transactions": "Access tarnsaction read operations."
            }
          }
        }
      }
    }
  },
...
}